ANY_SIMD: Add optimizations and refactoring in favor of speed using Return Value Optimization;

+++ period_search_optimization_simd/period_search/systeminfo.h
@@ -8,5 +8,5 @@ void getSystemInfo();
 #if defined(__arm__) || defined(__aarch64__) || defined(_M_ARM) || defined(_M_ARM64)
 void getCpuInfoByArch(std::ifstream &cpuinfo);
 #endif
-float getTotalSystemMemory();
+double getTotalSystemMemory();
